Implement Redis client utility and main operation in index.ts
